fix(company data): sonarl cloud duplicates issue
