+++ src/features/companyData/companyDataApiSlice.tsx
@@ -212,133 +212,6 @@ export enum SharingStateStatusType {
   Initial = 'Initial',
 }
 
-export const companyDataInitialData: CompanyDataType = {
-  externalId: '',
-  nameParts: [''],
-  identifiers: [
-    {
-      type: null,
-      value: null,
-      issuingBody: null,
-    },
-  ],
-  states: [
-    {
-      validFrom: null,
-      validTo: null,
-      type: 'ACTIVE',
-    },
-  ],
-  roles: ['SUPPLIER'],
-  isOwnCompanyData: false,
-  legalEntity: {
-    legalEntityBpn: '',
-    legalName: '',
-    shortName: '',
-    legalForm: '',
-    confidenceCriteria: {
-      sharedByOwner: false,
-      checkedByExternalDataSource: false,
-      numberOfSharingMembers: 0,
-      lastConfidenceCheckAt: '',
-      nextConfidenceCheckAt: '',
-      confidenceLevel: 0,
-    },
-    states: [
-      {
-        validFrom: null,
-        validTo: null,
-        type: 'ACTIVE',
-      },
-    ],
-  },
-  site: {
-    siteBpn: '',
-    name: '',
-    confidenceCriteria: {
-      sharedByOwner: false,
-      checkedByExternalDataSource: false,
-      numberOfSharingMembers: 0,
-      lastConfidenceCheckAt: '',
-      nextConfidenceCheckAt: '',
-      confidenceLevel: 0,
-    },
-    states: [
-      {
-        validFrom: null,
-        validTo: null,
-        type: 'ACTIVE',
-      },
-    ],
-  },
-  address: {
-    addressBpn: '',
-    name: '',
-    addressType: '',
-    physicalPostalAddress: {
-      geographicCoordinates: {
-        longitude: 0,
-        latitude: 0,
-        altitude: 0,
-      },
-      country: '',
-      administrativeAreaLevel1: '',
-      administrativeAreaLevel2: '',
-      administrativeAreaLevel3: '',
-      postalCode: '',
-      city: '',
-      district: '',
-      street: {
-        namePrefix: '',
-        additionalNamePrefix: '',
-        name: '',
-        nameSuffix: '',
-        additionalNameSuffix: '',
-        houseNumber: '',
-        houseNumberSupplement: '',
-        milestone: '',
-        direction: '',
-      },
-      companyPostalCode: '',
-      industrialZone: '',
-      building: '',
-      floor: '',
-      door: '',
-    },
-    alternativePostalAddress: {
-      geographicCoordinates: {
-        longitude: 0,
-        latitude: 0,
-        altitude: 0,
-      },
-      country: null,
-      administrativeAreaLevel1: null,
-      postalCode: null,
-      city: null,
-      deliveryServiceType: null,
-      deliveryServiceQualifier: null,
-      deliveryServiceNumber: null,
-    },
-    confidenceCriteria: {
-      sharedByOwner: true,
-      checkedByExternalDataSource: true,
-      numberOfSharingMembers: 0,
-      lastConfidenceCheckAt: null,
-      nextConfidenceCheckAt: null,
-      confidenceLevel: 0,
-    },
-    states: [
-      {
-        validFrom: null,
-        validTo: null,
-        type: 'ACTIVE',
-      },
-    ],
-  },
-  createdAt: '',
-  updatedAt: '',
-}
-
 export const apiSlice = createApi({
   reducerPath: 'rtk/companyData',
   baseQuery: fetchBaseQuery(apiBpdmGateQuery()),
